# docs/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "version": "1.0.0",
    "title": "Contacts api",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "description": "training"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    },
    {
      "url": null
    }
  ],
  "tags": null,
  "paths": {
    "/contacts": {
      "get": {
        "tags": [
          "Contacts"
        ],
        "summary": "Get all contacts",
        "operationId": "getAllContacts",
        "description": "Get all contacts with pagination, sorting and filtering",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "example": 1
            },
            "description": "Page number"
          },
          {
            "in": "query",
            "name": "perPage",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 10,
              "example": 10
            },
            "description": "Items per page"
          },
          {
            "in": "query",
            "name": "sortBy",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "name",
                "phoneNumber",
                "email",
                "contactType",
                "createdAt"
              ],
              "default": "createdAt",
              "example": "name"
            },
            "description": "Sort by field"
          },
          {
            "in": "query",
            "name": "sortOrder",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ],
              "default": "desc",
              "example": "asc"
            },
            "description": "Sort order"
          },
          {
            "in": "query",
            "name": "contactType",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "work",
                "home",
                "personal"
              ],
              "example": "personal"
            },
            "description": "Filter by contact type"
          },
          {
            "in": "query",
            "name": "isFavourite",
            "required": false,
            "schema": {
              "type": "boolean",
              "example": true
            },
            "description": "Filter by favourite status"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully get contacts",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully found contacts!"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "contacts": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/contact"
                          }
                        },
                        "pagination": {
                          "type": "object",
                          "properties": {
                            "page": {
                              "type": "integer",
                              "example": 1
                            },
                            "perPage": {
                              "type": "integer",
                              "example": 10
                            },
                            "total": {
                              "type": "integer",
                              "example": 25
                            },
                            "totalPages": {
                              "type": "integer",
                              "example": 3
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      },
      "post": {
        "tags": [
          "Contacts"
        ],
        "summary": "Create contact",
        "operationId": "createContact",
        "description": "Create a new contact",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "phoneNumber",
                  "contactType"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "Contact name",
                    "minLength": 3,
                    "maxLength": 20,
                    "example": "John Doe"
                  },
                  "phoneNumber": {
                    "type": "string",
                    "description": "Contact phone number",
                    "minLength": 3,
                    "maxLength": 20,
                    "example": 380991234567
                  },
                  "email": {
                    "type": "string",
                    "description": "Contact email",
                    "minLength": 3,
                    "maxLength": 20,
                    "example": "john@example.com"
                  },
                  "isFavourite": {
                    "type": "boolean",
                    "description": "Is contact favourite",
                    "example": false,
                    "default": false
                  },
                  "contactType": {
                    "type": "string",
                    "description": "Contact type",
                    "enum": [
                      "work",
                      "home",
                      "personal"
                    ],
                    "example": "personal",
                    "default": "personal"
                  },
                  "photo": {
                    "type": "string",
                    "format": "binary",
                    "description": "Contact photo"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Contact created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 201
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully created a contact!"
                    },
                    "data": {
                      "$ref": "#/components/schemas/contact"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
      "contact": {
        "type": "object",
        "required": [
          "_id",
          "name",
          "phoneNumber",
          "contactType",
          "userId",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "_id": {
            "type": "string",
            "description": "Contact id",
            "example": "507f191e810c19729de860ea"
          },
          "name": {
            "type": "string",
            "description": "Contact name",
            "example": "John Doe"
          },
          "phoneNumber": {
            "type": "string",
            "description": "Contact phone number",
            "example": 380991234567
          },
          "email": {
            "type": "string",
            "description": "Contact email",
            "example": "john@example.com"
          },
          "isFavourite": {
            "type": "boolean",
            "description": "Is contact favourite",
            "example": false,
            "default": false
          },
          "contactType": {
            "type": "string",
            "description": "Contact type",
            "enum": [
              "work",
              "home",
              "personal"
            ],
            "example": "personal"
          },
          "userId": {
            "type": "string",
            "description": "Contact owner id",
            "example": "6592008029c8c3e4dc76256c"
          },
          "photo": {
            "type": "string",
            "format": "url",
            "description": "Contact photo url",
            "example": "https://res.cloudinary.com/dha4xvgfi/image/upload/v1749060633/mef8zwaq73iedpo4owut.png"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Contact created at",
            "example": "2025-03-27T19:43:21.686+00:00"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Contact updated at",
            "example": "2025-03-28T19:43:21.686+00:00"
          }
        }
      }
    },
    "responses": {
      "401": {
        "description": "Unauthorized Error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message",
                "data"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 401
                },
                "message": {
                  "type": "string",
                  "example": "Unauthorized Error"
                },
                "data": {
                  "type": "object",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Access token not provided"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "500": {
        "description": "Internal Server Error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 500
                },
                "message": {
                  "type": "string",
                  "example": "Internal Server Error"
                }
              }
            }
          }
        }
      }
    }
  },
  "/contacts/{contactId}": {
    "get": {
      "$ref": "../swagger/paths/contacts/{contactId}/get.yaml"
    },
    "delete": {
      "$ref": "../swagger/paths/contacts/{contactId}/delete.yaml"
    },
    "patch": {
      "$ref": "../swagger/paths/contacts/{contactId}/patch.yaml"
    }
  },
  "/auth/register": {
    "post": {
      "$ref": "../swagger/paths/auth/register.yaml"
    }
  },
  "/auth/login": {
    "post": {
      "$ref": "../swagger/paths/auth/login.yaml"
    }
  },
  "/auth/send-reset-email": {
    "post": {
      "$ref": "../swagger/paths/auth/send-reset-email.yaml"
    }
  },
  "/auth/reset-pwd": {
    "post": {
      "$ref": "../swagger/paths/auth/reset-pwd.yaml"
    }
  },
  "/auth/refresh": {
    "post": {
      "$ref": "../swagger/paths/auth/refresh.yaml"
    }
  },
  "/auth/logout": {
    "post": {
      "$ref": "../swagger/paths/auth/logout.yaml"
    }
  },
  "/auth/get-oauth-url": {
    "get": {
      "$ref": "../swagger/paths/auth/get-oauth-url.yaml"
    }
  },
  "/auth/confirm-oauth": {
    "post": {
      "$ref": "../swagger/paths/auth/confirm-oauth.yaml"
    }
  },
  "schemas": {
    "contact": {
      "$ref": "../swagger/components/schemas/contact.yaml"
    },
    "user": {
      "$ref": "../swagger/components/schemas/user.yaml"
    }
  },
  "responses": {
    "401": {
      "$ref": "../swagger/components/responses/401.yaml"
    },
    "500": {
      "$ref": "../swagger/components/responses/500.yaml"
    }
  }
}